requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
pinecone[grpc]
requests
pydantic
